functions/consume_iam_anomalies_func/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
google-api-core[grpc]==1.30.0
    # via
    #   google-cloud-core
    #   google-cloud-datastore
google-auth==1.31.0
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-core==1.6.0
    # via google-cloud-datastore
google-cloud-datastore==2.1.3
    # via -r requirements.in
googleapis-common-protos==1.53.0
    # via google-api-core
grpcio==1.38.0
    # via google-api-core
idna==2.10
    # via requests
libcst==0.3.19
    # via google-cloud-datastore
mypy-extensions==0.4.3
    # via typing-inspect
packaging==20.9
    # via google-api-core
proto-plus==1.18.1
    # via google-cloud-datastore
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via google-api-core
pyyaml==5.4.1
    # via libcst
requests==2.25.1
    # via google-api-core
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-core
    #   grpcio
    #   protobuf
typing-extensions==3.10.0.0
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
urllib3==1.26.5
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
